Move PATH testutils to testutil crate
